# openapi/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Vonk Forge Catalog API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/recipes": {
      "get": {
        "operationId": "list_recipes_v1_recipes_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "publisher",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 63,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publisher"
            }
          },
          {
            "in": "query",
            "name": "official",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Official"
            }
          },
          {
            "in": "query",
            "name": "runtime",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Runtime"
            }
          },
          {
            "in": "query",
            "name": "workload_family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workload Family"
            }
          },
          {
            "in": "query",
            "name": "capability",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Capability"
            }
          },
          {
            "in": "query",
            "name": "topology",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(single|gang)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topology"
            }
          },
          {
            "in": "query",
            "name": "tested_node_count",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 16,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tested Node Count"
            }
          },
          {
            "in": "query",
            "name": "min_nodes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 16,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Nodes"
            }
          },
          {
            "in": "query",
            "name": "max_nodes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 16,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Nodes"
            }
          },
          {
            "in": "query",
            "name": "max_memory_bytes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Memory Bytes"
            }
          },
          {
            "in": "query",
            "name": "max_installed_bytes",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Installed Bytes"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "newest",
              "pattern": "^(newest|title|disk|memory)$",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Recipes V1 Recipes Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Recipes"
      }
    },
    "/v1/recipes/{publisher}/{slug}": {
      "get": {
        "operationId": "recipe_detail_v1_recipes__publisher___slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "publisher",
            "required": true,
            "schema": {
              "title": "Publisher",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Recipe Detail V1 Recipes  Publisher   Slug  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recipe Detail"
      }
    },
    "/v1/recipes/{publisher}/{slug}/revisions/{revision_number}": {
      "get": {
        "operationId": "recipe_revision_v1_recipes__publisher___slug__revisions__revision_number__get",
        "parameters": [
          {
            "in": "path",
            "name": "publisher",
            "required": true,
            "schema": {
              "title": "Publisher",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_number",
            "required": true,
            "schema": {
              "title": "Revision Number",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recipe Revision"
      }
    },
    "/v1/schemas/recipe/v1": {
      "get": {
        "operationId": "recipe_schema_v1_schemas_recipe_v1_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Recipe Schema V1 Schemas Recipe V1 Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Recipe Schema"
      }
    }
  }
}